+++ src/main/java/com/example/auctionmarket/common/config/WebSocketConfig.java
@@ -1,11 +1,11 @@
-package com.example.auctionmarket.common.websocket;
+package com.example.auctionmarket.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class WebConfig {
+public class WebSocketConfig {
 
     @Bean
     public RestTemplate restTemplate() {
