--- src/main/java/com/example/auctionmarket/common/redis/RedisConfig.java
@@ -1,43 +1,0 @@
-//package com.example.auctionmarket.common.redis;
-//
-//import org.springframework.beans.factory.annotation.Value;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.data.redis.connection.RedisPassword;
-//import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-//import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-//import org.springframework.data.redis.core.RedisTemplate;
-//import org.springframework.data.redis.serializer.StringRedisSerializer;
-//
-//@Configuration("redisConfigB")
-//public class RedisConfig {
-//    @Value("${spring.data.redis.host}")
-//    private String host;
-//
-//    @Value("${spring.data.redis.port}")
-//    private int port;
-//
-//    @Value("${spring.data.redis.password}")
-//    private String password;
-//
-//    @Bean
-//    public LettuceConnectionFactory redisConnectionFactory() { // Lettuce라는 라이브러리
-//
-//        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration(host, port);
-//        config.setPassword(RedisPassword.of(password)); // ✅ 비밀번호 주입
-//        return new LettuceConnectionFactory(config);
-//
-//    }
-//    @Bean
-//    public RedisTemplate<String, Object> redisTemplate() {
-//        RedisTemplate<String, Object> template = new RedisTemplate<>();
-//
-//        template.setConnectionFactory(redisConnectionFactory());
-//        template.setKeySerializer(new StringRedisSerializer());
-//        template.setValueSerializer(new StringRedisSerializer());
-//
-//        return template;
-//    }
-//
-//}
-
