+++ src/main/java/com/example/auctionmarket/domain/product/entity/Product.java
@@ -1,16 +1,29 @@
 package com.example.auctionmarket.domain.product.entity;
 
-import com.example.auctionmarket.common.entity.TimeStamped;
+import java.util.ArrayList;
+import java.util.List;
+
+import com.example.auctionmarket.common.entity.BaseEntity;
 import com.example.auctionmarket.domain.product.enums.ProductCategory;
 import com.example.auctionmarket.domain.product.enums.SoldStatus;
 import com.example.auctionmarket.domain.productimage.entity.ProductImage;
 import com.example.auctionmarket.domain.user.entity.User;
-import jakarta.persistence.*;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Index;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
 @Entity
@@ -18,7 +31,7 @@ import java.util.List;
 @Table(name = "product", indexes = {
         @Index(name = "idx_product_name", columnList = "category, productName")
 })
-public class Product extends TimeStamped {
+public class Product extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
