+++ src/main/java/com/example/auctionmarket/domain/productimage/entity/ProductImage.java
@@ -1,15 +1,23 @@
 package com.example.auctionmarket.domain.productimage.entity;
 
-import com.example.auctionmarket.common.entity.TimeStamped;
+import com.example.auctionmarket.common.entity.BaseEntity;
 import com.example.auctionmarket.domain.product.entity.Product;
-import jakarta.persistence.*;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
 @Entity(name = "product_image")
-public class ProductImage extends TimeStamped {
+public class ProductImage extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
