--- src/main/java/com/example/auctionmarket/domain/user/entity/TimeStamped.java
@@ -1,30 +1,0 @@
-package com.example.auctionmarket.domain.user.entity;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
-import lombok.Getter;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public class TimeStamped {
-
-	@CreatedDate
-	@Column(updatable = false)
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime createdAt;
-
-	@LastModifiedDate
-	@Column
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime updatedAt;
-}
