--- src/main/java/com/example/auctionmarket/global/config/RedisCacheConfig.java
@@ -1,41 +1,0 @@
-//package com.example.auctionmarket.global.config;
-//
-//import org.springframework.cache.CacheManager;
-//import org.springframework.cache.annotation.EnableCaching;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.data.redis.cache.RedisCacheConfiguration;
-//import org.springframework.data.redis.cache.RedisCacheManager;
-//import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-//import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-//import org.springframework.data.redis.serializer.RedisSerializationContext;
-//import org.springframework.data.redis.serializer.StringRedisSerializer;
-//
-//import jakarta.annotation.PreDestroy;
-//import lombok.RequiredArgsConstructor;
-//
-//@Configuration
-//@EnableCaching
-//@RequiredArgsConstructor
-//public class RedisCacheConfig {
-//
-//	private final LettuceConnectionFactory lettuceConnectionFactory;
-//
-//	@Bean
-//	public CacheManager redisCache() {
-//		RedisCacheConfiguration redisCacheConfiguration = RedisCacheConfiguration.defaultCacheConfig()
-//			.serializeKeysWith(RedisSerializationContext.SerializationPair.fromSerializer(new StringRedisSerializer()))
-//			.serializeValuesWith(
-//				RedisSerializationContext.SerializationPair.fromSerializer(new GenericJackson2JsonRedisSerializer()));
-//
-//		return RedisCacheManager.RedisCacheManagerBuilder
-//			.fromConnectionFactory(lettuceConnectionFactory)
-//			.cacheDefaults(redisCacheConfiguration)
-//			.build();
-//	}
-//
-//	@PreDestroy
-//	public void destroy() {
-//		lettuceConnectionFactory.destroy();
-//	}
-//}
