--- src/main/java/com/example/auctionmarket/global/config/RedisConfig.java
@@ -1,54 +1,0 @@
-//package com.example.auctionmarket.global.config;
-//
-//import org.springframework.beans.factory.annotation.Value;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.core.io.ClassPathResource;
-//import org.springframework.data.redis.connection.RedisConnectionFactory;
-//import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-//import org.springframework.data.redis.core.RedisTemplate;
-//import org.springframework.data.redis.core.script.DefaultRedisScript;
-//import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-//import org.springframework.data.redis.serializer.StringRedisSerializer;
-//import org.springframework.scripting.support.ResourceScriptSource;
-//
-//@Configuration
-//public class RedisConfig {
-//	@Bean
-//	public DefaultRedisScript<Long> redisScript() {
-//		DefaultRedisScript<Long> redisScript = new DefaultRedisScript<>();
-//		redisScript.setScriptSource(new ResourceScriptSource(new ClassPathResource("scripts/reserve_seat.lua")));
-//		redisScript.setResultType(Long.class);
-//		return redisScript;
-//	}
-//
-//	@Value("${spring.data.redis.host}")
-//	private String host;
-//
-//	@Value("${spring.data.redis.port}")
-//	private int port;
-//
-//	private static final String REDISSON_HOST_PREFIX = "redis://";
-//
-//	// 오류로 추가한 것
-//	@Bean
-//	public RedisConnectionFactory redisConnectionFactory() {
-//		return new LettuceConnectionFactory(host, port);
-//	}
-//
-//	@Bean
-//	public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
-//		RedisTemplate<String, Object> template = new RedisTemplate<>();
-//		template.setConnectionFactory(redisConnectionFactory); // ConnectionFactory 설정
-//
-//		// Serializer 설정 (중요!)
-//		template.setKeySerializer(new StringRedisSerializer());
-//		template.setValueSerializer(new GenericJackson2JsonRedisSerializer()); // Value를 JSON으로
-//		template.setHashKeySerializer(new StringRedisSerializer());
-//		template.setHashValueSerializer(new GenericJackson2JsonRedisSerializer());
-//
-//		return template;
-//	}
-//
-//}
-
