--- src/test/java/com/example/auctionmarket/AuctionMarketApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.auctionmarket;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuctionMarketApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
